+++ test/date_time/is_same_test.dart
@@ -1,5 +1,5 @@
 import 'package:test/test.dart';
-import 'package:time_plus/date_time/is_same.dart';
+import 'package:time_plus/time_plus.dart';
 
 void main() {
   group('DateTime isSame extensions', () {